replaced ExecMode with proper UnrealTargetType
